Admission time admissions 2 hours before approaching the attention is to send an email notice
