+++ src/main/java/uz/dev/hmsproject/repository/AppointmentRepository.java
@@ -77,4 +77,7 @@ public interface AppointmentRepository extends JpaRepository<Appointment, Long> 
 
 
     List<Appointment> findAllByStatus(AppointmentStatus appointmentStatus);
+
+    List<Appointment> findAllByAppointmentDateTimeBetweenAndStatus(LocalDateTime from, LocalDateTime to, AppointmentStatus appointmentStatus);
+
 }
