+++ src/main/java/uz/dev/hmsproject/service/AppointmentServiceImpl.java
@@ -269,6 +269,7 @@ public class AppointmentServiceImpl implements AppointmentService {
     }
 
     @Override
+    @Transactional
     public void changeStatus() {
         List<Appointment> appointments = appointmentRepository.findAllByStatus(AppointmentStatus.SCHEDULED);
